+++ src/COM/ConfigurationManager.cpp
@@ -94,7 +94,7 @@ BSTR ConfigurationManager::AppSettings(const BSTR key, PCWSTR defaultValue) noex
 	//note: GetValueAt returns by reference, and does not Copy()	
 	if (found != _map.end() && found->first == key) 
 	{
-		return ::SysAllocStringLen(found->second.c_str(), found->second.length());
+		return ::SysAllocStringLen(found->second.c_str(), static_cast<UINT>( found->second.length()));
 	}
 	else if (defaultValue != nullptr)
 	{
